+++ CHANGELOG.rst
@@ -6,6 +6,16 @@ snovault
 Change Log
 ----------
 
+11.22.0
+=======
+* 2024-09-03/dmichaels
+  - Fix in snovault/tests/elasticsearch_fixture.py (use only for local/dev deploy) for
+    strange (new as of 2024-09-02) behavior where it was hanging on startup during
+    ElasticSearch index mapping creation, related to ElasticSearch logging output,
+    and the way we were using subprocess.Popen and reading the subprocess output; 
+    more correct way is to inherit stdout/stderr of the partent.
+
+
 11.21.1
 =======